+++ selfdrive/modeld/dmonitoringmodeld.cc
@@ -30,7 +30,7 @@ void run_model(DMonitoringModelState &model, VisionIpcClient &vipc_client) {
     // send dm packet
     dmonitoring_publish(pm, extra.frame_id, res, (t2 - t1) / 1000.0, model.output);
 
-    LOGD("dmonitoring process: %.2fms, from last %.2fms", t2 - t1, t1 - last);
+    //printf("dmonitoring process: %.2fms, from last %.2fms\n", t2 - t1, t1 - last);
     last = t1;
   }
 }