+++ selfdrive/ui/qt/offroad/onboarding.cc
@@ -1,30 +1,17 @@
 #include <QLabel>
 #include <QString>
+#include <QScroller>
+#include <QScrollBar>
 #include <QPushButton>
 #include <QGridLayout>
 #include <QVBoxLayout>
 #include <QDesktopWidget>
+#include <QPainter>
 
 #include "common/params.h"
 #include "onboarding.hpp"
 #include "home.hpp"
 #include "util.h"
-
-
-QLabel * title_label(QString text) {
-  QLabel *l = new QLabel(text);
-  l->setStyleSheet(R"(
-    font-size: 100px;
-    font-weight: 400;
-  )");
-  return l;
-}
-
-QWidget * layout2Widget(QLayout* l){
-  QWidget *q = new QWidget;
-  q->setLayout(l);
-  return q;
-}
 
 
 void TrainingGuide::mouseReleaseEvent(QMouseEvent *e) {
@@ -33,61 +20,59 @@ void TrainingGuide::mouseReleaseEvent(QMouseEvent *e) {
   int mousey = e->y();
 
   // Check for restart
-  if (currentIndex == boundingBox.size()-1) {
-    if (1050 <= mousex && mousex <= 1500 && 773 <= mousey && mousey <= 954){
-      slayout->setCurrentIndex(0);
-      currentIndex = 0;
-      return;
-    }
+  if (currentIndex == (boundingBox.size() - 1) && 1050 <= mousex && mousex <= 1500 &&
+      773 <= mousey && mousey <= 954) {
+    currentIndex = 0;
+  } else if (boundingBox[currentIndex][0] <= mousex && mousex <= boundingBox[currentIndex][1] &&
+             boundingBox[currentIndex][2] <= mousey && mousey <= boundingBox[currentIndex][3]) {
+    currentIndex += 1;
   }
 
-  if (boundingBox[currentIndex][0] <= mousex && mousex <= boundingBox[currentIndex][1] && boundingBox[currentIndex][2] <= mousey && mousey <= boundingBox[currentIndex][3]) {
-    slayout->setCurrentIndex(++currentIndex);
-  }
   if (currentIndex >= boundingBox.size()) {
     emit completedTraining();
     return;
+  } else {
+    image.load("../assets/training/step" + QString::number(currentIndex) + ".jpg");
+    update();
   }
 }
 
 TrainingGuide::TrainingGuide(QWidget* parent) {
-  QHBoxLayout* hlayout = new QHBoxLayout;
+  image.load("../assets/training/step0.jpg");
+}
 
-  slayout = new QStackedLayout(this);
-  for (int i = 0; i < boundingBox.size(); i++) {
-    QWidget* w = new QWidget;
-    w->setStyleSheet(".QWidget {background-image: url(../assets/training/step" + QString::number(i) + ".jpg);}");
-    w->setFixedSize(1620, 1080);
-    slayout->addWidget(w);
-  }
+void TrainingGuide::paintEvent(QPaintEvent *event) {
+  QPainter painter(this);
 
-  QWidget* sw = layout2Widget(slayout);
-  hlayout->addWidget(sw, 1, Qt::AlignCenter);
-  setLayout(hlayout);
-  setStyleSheet(R"(
-    background-color: #072339;
-  )");
+  QRect devRect(0, 0, painter.device()->width(), painter.device()->height());
+  QBrush bgBrush("#072339");
+  painter.fillRect(devRect, bgBrush);
+
+  QRect rect(image.rect());
+  rect.moveCenter(devRect.center());
+  painter.drawImage(rect.topLeft(), image);
 }
 
 
 QWidget* OnboardingWindow::terms_screen() {
   QVBoxLayout *main_layout = new QVBoxLayout;
-  main_layout->setContentsMargins(40, 0, 40, 0);
+  main_layout->setContentsMargins(40, 20, 40, 20);
 
-#ifndef QCOM
-  view = new QWebEngineView(this);
-  view->settings()->setAttribute(QWebEngineSettings::ShowScrollBars, false);
-  QString html = QString::fromStdString(util::read_file("../assets/offroad/tc.html"));
-  view->setHtml(html);
-  main_layout->addWidget(view);
+  QString terms_html = QString::fromStdString(util::read_file("../assets/offroad/tc.html"));
+  terms_text = new QTextEdit();
+  terms_text->setReadOnly(true);
+  terms_text->setTextInteractionFlags(Qt::NoTextInteraction);
+  terms_text->setHtml(terms_html);
+  main_layout->addWidget(terms_text);
 
-  QObject::connect(view->page(), SIGNAL(scrollPositionChanged(QPointF)), this, SLOT(scrollPositionChanged(QPointF)));
-#endif
-
+  // TODO: add decline page
   QHBoxLayout* buttons = new QHBoxLayout;
+  main_layout->addLayout(buttons);
+
   buttons->addWidget(new QPushButton("Decline"));
   buttons->addSpacing(50);
-  accept_btn = new QPushButton("Scroll to accept");
+
+  QPushButton *accept_btn = new QPushButton("Scroll to accept");
   accept_btn->setEnabled(false);
   buttons->addWidget(accept_btn);
   QObject::connect(accept_btn, &QPushButton::released, [=]() {
@@ -95,15 +80,25 @@ QWidget* OnboardingWindow::terms_screen() {
     updateActiveScreen();
   });
 
-  QWidget* w = layout2Widget(buttons);
-  w->setFixedHeight(200);
-  main_layout->addWidget(w);
+  // TODO: tune the scrolling
+  auto sb = terms_text->verticalScrollBar();
+  QScroller::grabGesture(terms_text, QScroller::TouchGesture);
+  terms_text->setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
+
+  QObject::connect(sb, &QScrollBar::valueChanged, [sb, accept_btn]() {
+    if (sb->value() == sb->maximum()){
+      accept_btn->setText("Accept");
+      accept_btn->setEnabled(true);
+    }
+  });
 
   QWidget *widget = new QWidget;
   widget->setLayout(main_layout);
   widget->setStyleSheet(R"(
-    QPushButton {
+    * {
       font-size: 50px;
+    }
+    QPushButton {
       padding: 50px;
       border-radius: 10px;
       background-color: #292929;
@@ -114,10 +109,9 @@ QWidget* OnboardingWindow::terms_screen() {
 }
 
 void OnboardingWindow::updateActiveScreen() {
-  Params params = Params();
-
   bool accepted_terms = params.get("HasAcceptedTerms", false).compare(current_terms_version) == 0;
   bool training_done = params.get("CompletedTrainingVersion", false).compare(current_training_version) == 0;
+
   if (!accepted_terms) {
     setCurrentIndex(0);
   } else if (!training_done) {
@@ -128,16 +122,10 @@ void OnboardingWindow::updateActiveScreen() {
 }
 
 OnboardingWindow::OnboardingWindow(QWidget *parent) : QStackedWidget(parent) {
-  Params params = Params();
+  params = Params();
   current_terms_version = params.get("TermsVersion", false);
   current_training_version = params.get("TrainingVersion", false);
-  bool accepted_terms = params.get("HasAcceptedTerms", false).compare(current_terms_version) == 0;
-  bool training_done = params.get("CompletedTrainingVersion", false).compare(current_training_version) == 0;
-  
-  //Don't initialize widgets unless neccesary. 
-  if (accepted_terms && training_done) {
-    return;
-  }
+
   addWidget(terms_screen());
 
   TrainingGuide* tr = new TrainingGuide(this);
@@ -164,13 +152,4 @@ OnboardingWindow::OnboardingWindow(QWidget *parent) : QStackedWidget(parent) {
   )");
 
   updateActiveScreen();
-}
-
-void OnboardingWindow::scrollPositionChanged(QPointF position){
-#ifndef QCOM
-  if (position.y() > view->page()->contentsSize().height() - 1000){
-    accept_btn->setEnabled(true);
-    accept_btn->setText("Accept");
-  }
-#endif
 }