+++ selfdrive/ui/qt/offroad/settings.hpp
@@ -3,28 +3,31 @@
 #include <QWidget>
 #include <QFrame>
 #include <QTimer>
+#include <QLabel>
 #include <QPushButton>
 #include <QButtonGroup>
-#include <QStackedLayout>
+#include <QScrollArea>
+#include <QStackedWidget>
 
-// *** settings widgets ***
+#include "selfdrive/ui/qt/widgets/controls.hpp"
 
-class ParamsToggle : public QFrame {
+// ********** settings window + top-level panels **********
+
+class DevicePanel : public QWidget {
   Q_OBJECT
-
 public:
-  explicit ParamsToggle(QString param, QString title, QString description,
-                        QString icon, QWidget *parent = 0);
-
-private:
-  QString param;
-
-public slots:
-  void checkboxClicked(int state);
+  explicit DevicePanel(QWidget* parent = nullptr);
 };
 
+class DeveloperPanel : public QFrame {
+  Q_OBJECT
+public:
+  explicit DeveloperPanel(QWidget* parent = nullptr);
 
-// *** settings window ***
+protected:
+  void showEvent(QShowEvent *event) override;
+  QList<LabelControl *> labels;
+};
 
 class SettingsWindow : public QFrame {
   Q_OBJECT
@@ -34,16 +37,11 @@ public:
 
 signals:
   void closeSettings();
-  void sidebarPressed();
 
 private:
   QPushButton *sidebar_alert_widget;
   QWidget *sidebar_widget;
-  std::map<QString, QWidget *> panels;
   QButtonGroup *nav_btns;
-  QStackedLayout *panel_layout;
-  QFrame* panel_frame;
-
-public slots:
-  void setActivePanel();
+  QStackedWidget *panel_widget;
+  QScrollArea *panel_frame;
 };