+++ selfdrive/ui/qt/qt_sound.hpp
@@ -9,6 +9,7 @@ public:
   bool play(AudibleAlert alert);
   void stop();
   void setVolume(int volume);
+  float volume = 0;
 
 private:
   std::map<AudibleAlert, QSoundEffect> sounds;