+++ tools/sim/build_container.sh
@@ -3,8 +3,8 @@
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR/../../
 
-docker pull commaai/openpilot-base:latest
+docker pull ghcr.io/commaai/openpilot-base:latest
 docker build \
-  --cache-from commaai/openpilot-sim:latest \
-  -t commaai/openpilot-sim:latest \
+  --cache-from ghcr.io/commaai/openpilot-sim:latest \
+  -t ghcr.io/commaai/openpilot-sim:latest \
   -f tools/sim/Dockerfile.sim .