+++ RELEASES.md
@@ -1,5 +1,8 @@
 Version 0.8.3 (2021-XX-XX)
 ========================
+ * NEOS update: update packages and support for new UI
+ * New offroad UI based on Qt
+ * Default SSH key only used for setup
  * Kia Ceed 2019 support thanks to ZanZaD13!
  * Kia Seltos 2021 support thanks to speedking456!